Fix blank screen and unserialisable Tag enum
Two runtime failures that neither the compiler nor 37 passing tests caught,
because both only appear when Unity's asset pipeline and render loop actually
run.

Tag is now uint, not ulong (ADR-022)
  Unity's serializer rejects 64-bit-backed enums outright:
    Unsupported enum type 'Overrun.Core.Tag' used for field 'Tags'
  The field simply does not serialise, so every WeaponDefinition and
  EnemyDefinition lost its tags and the tag-filtered modifier system had nothing
  left to filter on — taking the whole data-driven definition layer with it.
  32 flags max, 21 used. If we ever approach the ceiling the fix is a
  serializable two-uint struct, NOT widening the enum back.

Lobby camera
  There were zero cameras in any scene. Player cameras live on the rig prefab,
  which PlayerInputManager only instantiates once a device presses a button, so
  the game opened to a genuinely blank screen. LocalRigs now carries a
  LobbyCamera that LocalPlayerJoinHandler disables on first join and restores
  when the last local player leaves.

Bootstrap scripts no longer clobber each other
  VS001Bootstrap rebuilt PlayerPawn.prefab unconditionally, silently destroying
  the WeaponRuntime and weapon reference that VS001Combat had added — producing
  a pawn that compiles, spawns and cannot shoot. Prefab builders now preserve
  existing assets, and a chained "Overrun/Run VS001 (All)" entry point runs the
  passes in the required order.

Docs
  GAMEPLAY_SYSTEMS.md §2 corrected (it specified `: ulong`), ADR-022 added.
  Same lesson as ADR-021: engine serialisation constraints are invisible to both
  the compiler and headless tests.

37 EditMode tests still passing.

+++ Assets/Overrun.Core/Tag.cs
@@ -10,38 +10,46 @@ namespace Overrun.Core
     /// "Lightning Rounds" (Projectile|Elemental|Shock), "Conductive Blood" (Shock|Status)
     /// and "Overcharge" (Shock|Critical) form an electrical build purely because each
     /// filters on Shock. No code knows those three are related.
+    ///
+    /// UNDERLYING TYPE IS uint, NOT ulong — 32 tags maximum.
+    /// Unity's serializer rejects enums with a 64-bit backing type outright:
+    ///     "Unsupported enum type 'Overrun.Core.Tag' used for field 'Tags'"
+    /// A ulong-backed Tag cannot appear on a ScriptableObject field, which would break the
+    /// entire data-driven definition layer. 21 of 32 bits are used; if we ever approach
+    /// the ceiling the fix is a serializable two-uint struct with implicit conversion to a
+    /// 64-bit runtime value, NOT widening this enum.
     /// </summary>
     [Flags]
-    public enum Tag : ulong
+    public enum Tag : uint
     {
-        None = 0UL,
+        None = 0u,
 
         // --- Source
-        Weapon     = 1UL << 0,
-        Projectile = 1UL << 1,
-        Hitscan    = 1UL << 2,
-        Melee      = 1UL << 3,
-        Ability    = 1UL << 4,
-        Companion  = 1UL << 5,
-        Trap       = 1UL << 6,
-        Explosion  = 1UL << 7,
+        Weapon     = 1u << 0,
+        Projectile = 1u << 1,
+        Hitscan    = 1u << 2,
+        Melee      = 1u << 3,
+        Ability    = 1u << 4,
+        Companion  = 1u << 5,
+        Trap       = 1u << 6,
+        Explosion  = 1u << 7,
 
         // --- Elements
-        Shock = 1UL << 8,
-        Fire  = 1UL << 9,
-        Frost = 1UL << 10,
-        Toxic = 1UL << 11,
-        Void  = 1UL << 12,
+        Shock = 1u << 8,
+        Fire  = 1u << 9,
+        Frost = 1u << 10,
+        Toxic = 1u << 11,
+        Void  = 1u << 12,
 
         // --- Roles
-        Critical  = 1UL << 16,
-        Status    = 1UL << 17,
-        Defense   = 1UL << 18,
-        Mobility  = 1UL << 19,
-        Economy   = 1UL << 20,
-        Support   = 1UL << 21,
-        Elemental = 1UL << 22,
-        Area      = 1UL << 23,
+        Critical  = 1u << 16,
+        Status    = 1u << 17,
+        Defense   = 1u << 18,
+        Mobility  = 1u << 19,
+        Economy   = 1u << 20,
+        Support   = 1u << 21,
+        Elemental = 1u << 22,
+        Area      = 1u << 23,
 
         AnyElement = Shock | Fire | Frost | Toxic | Void
     }
